analysis/requirements.txt
requests
python-dotenv
pandas
numpy
matplotlib
tensorflow
scikit-learn
ipython
pydot